django_complete/requirements.txt
Django==4.2.7
langchain==0.1.20
langchain-google-genai==1.0.10
langchain-community==0.0.38
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.3
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.1.0
